Move GitHub discovery code under pkg/discovery
The discovery worker, scanner, and synthesis logic belong in a
discovery namespace rather than a generic integration folder.
