+++ pkg/discovery/github/connector_http.go
