+++ pkg/discovery/github/prompts/synthesis.txt.tmpl
