+++ pkg/discovery/github/scopes.go
