+++ pkg/discovery/github/scopes_test.go
