+++ pkg/discovery/github/service_test.go
