+++ pkg/discovery/github/synthesis_agent.go
