+++ pkg/discovery/github/types.go
