+++ pkg/discovery/github/worker.go
