+++ pkg/server/api/api.go
@@ -31,11 +31,11 @@ import (
 	"go.probo.inc/probo/pkg/connector"
 	"go.probo.inc/probo/pkg/connector/provider"
 	"go.probo.inc/probo/pkg/cookiebanner"
+	ghdiscovery "go.probo.inc/probo/pkg/discovery/github"
 	"go.probo.inc/probo/pkg/esign"
 	"go.probo.inc/probo/pkg/filemanager"
 	"go.probo.inc/probo/pkg/geoloc"
 	"go.probo.inc/probo/pkg/iam"
-	ghintegration "go.probo.inc/probo/pkg/integration/github"
 	"go.probo.inc/probo/pkg/mailman"
 	"go.probo.inc/probo/pkg/probo"
 	"go.probo.inc/probo/pkg/resourcealias"
@@ -71,7 +71,7 @@ type (
 		CookieBanner      *cookiebanner.Service
 		Geoloc            *geoloc.Service
 		ThirdParty        *thirdparty.Service
-		GitHubDiscovery   *ghintegration.Service
+		GitHubDiscovery   *ghdiscovery.Service
 		RiskManagement    *riskmanagement.Service
 		Cookie            securecookie.Config
 		TokenSecret       string
